Outline of code
I created edm_bias which will be the main bias class. Eventually I'll
split it into a pairwise and absolute coordinate. There are two grid
utilities, one for vanilla grids and one for gaussians. That's where
most of the work will be. The edm.cpp file will be what hooks into
lammps.