Changed the code to be very careful about bias energy
I've changed many pieces of the code so that it adds bias at a given
rate, independent of hill_density and mpi node number. This requires
being a little careful so that the sampling of hills isn't biased. If
you just stopped adding hills at a given step, that would bias the
sampling. This proved to be a little tricky for the pairwise case,
where the number of possible hills (pairs) is unknown. I did this by
stopping when the bias amount is reached and then utilizing the
pairs added afterwards in the next step.